requirements.txt
bertopic==0.14.1
nltk==3.8.1
numpy==1.23.5
pymongo==4.3.3
pytz==2023.3
river==0.16.0
tensorflow==2.10.0
torch==1.13.1+cu117
transformers==4.26.1
tweepy==4.13.0
umap_learn==0.5.3
chromadb>=0.4.0,<0.5.0
sentence-transformers>=2.0.0,<3.0.0
ollama>=0.1.0,<0.2.0
